doc: Use correct line break to fix build
Escaping the backslash we want to use for a line break is a really bad idea.
Introduced in f126afbba21af39c493f21424b653bb5c2e0d58f to support a crippled
echo without new line option.

The linebreak was not detected and thus the first image name used as a target.
Broke our nightly efl build with make distcheck and made me really unhappy to
while searching for the cause.