ecore_x_selection: do not skip any any atoms
Summary:
i dont know why we skipped the first two atoms, but right now, if a
application is only providing one single target, we would crash.
With this we might copy a few atoms more. However, these atoms do not
matter, as we skip those, that we cannot understand

Depends on D11193

Reviewers: cedric, zmike, raster, segfaultxavi, stefan_schmidt

Subscribers: #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D11194