eina_inline_thread_win32.x: Fix "user after free" bug
The `thread_func` variable is created in the caller's stack and used in
the created thread. If the caller's return before the thread is
executed, its stack frame is freed and `thread_func` is no longer valid.

We now create `thread_func` inside the heap so it remains valid after
the caller returns, and the entry point in the new thread is responsible
to free it.