build: Use meson subprojects (wrap-dependency) as fallback for dependencies
Summary:
Meson has a nice feature for dependencies called ["Wrap
Dependency"](https://mesonbuild.com/Wrap-dependency-system-manual.html), which
downloads dependency source-code and builds it using Meson. This way, it
becomes easier to assume library support for some dependencies such as `zlib`,
`libjpeg`, `libcheck` and etc. natively on some platforms such as Windows.

If the dependency is found, then the build-system will get things from it. It
it is not found, then the subproject will be used.

This is being used to help with T8760.

Test Plan:
In a Linux environment, ensure `libcheck` is uninstalled or
unreachable (as other deps with wraps added here will probably be tied with the
system, UI's packages, etc.) and then build EFL and run tests (e.g. `meson test
-C build`). Meson should download libcheck's (or any other wrap that is not
installed or is unreachable) source code and then build them when trying to
build EFL.

Reviewers: bu5hm4n, vtorri

Subscribers: cedric, #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D12034