fix: The -I should be passed to the linker.
The meson has a buildin argument for `declare_dependency()` called
`include_directories` but to use it the path should NOT be an absolute path.
To be able to use absolute path here we need to use 'I' include flag.
The equivalent for `include_directories` here is `link_args` and not
`compile_args`.