build: Properly handle --verbose arg
In `configure.bat`, the user may want to add custom flags to meson
calling:

```console
$ configure <extra args to meson>
```

Meanwhile, a "verbose" mode was also available for configure, which
would send `-v` flag to the `clang-cl` compiler (making it always output
its command line, so there would be no need to check every `.rsp` to see
compiler flags).

Previously, verbose output from configure only had effect if `--verbose`
was the **first** argument. Not only that, we had a hidden bug on it:
- Since `%MESONFLAGS_EXTRA%` was used (instead of `!MESONFLAGS_EXTRA!`),
  no flags were sent to meson from `configure` command line (because of
  how delayed expansion works on Windows);
- By fixing the usage of  `!` instead of `%`, since there was no
  filter on extra args sent to meson, the `--verbose` flag was also
  being sent, which would make `meson` fail due to not having a
  `--verbose` flag.

- Fix sending extra flags to meson replacing `%MESONFLAGS_EXTRA%`
  with `!MESONFLAGS_EXTRA!`;
- Filter `configure.bat`-specific args by not sending them to
  meson.