Eet: Make eet_test_cache test-case work
This test case had 2 problems:

`/tmp` doesn't exists in windows and `mkstemp` expects that all path
give already exists (besids the file it will create), causing a fail:
```
74:    fail_if(-1 == (tmpfd = mkstemp(file)));
```
Here this is solved by using [eina_environment_tmp_get()](https://github.com/expertisesolutions/efl/blob/561410306ee0d00763af2b5d2d0f360de782270c/src/lib/eina/eina_util.c#L97),
which is an architecture-independent function that returns the path to
the temporary directory, as the base of the path.

`eet_close` is a postponed close so when arriving at `unlink` the file
may not be closed yet. On linux it isn't a problem, as when `unlink`ing an
open file it returns `0` and waits until all references to the file to
be closed for its removal.
But on windows you while `unlink`ing an open file `-1` is returnged,  causing a fail:
´´´
102:  fail_if(unlink(file) != 0);
´´´
With `unlink` setting `errno` to `13`(`strerror(13)` = `Permission denied`).

As this is a test about cleaning cache it seams reasonable to force
cache cleaning at the end instead of closing the file and thus assuring
that `unlink` will return `0` on every architecture.