Eet: Resolve eet_test_image test case
Like `eet_test_cache`, this test case had 2 problems:

`/tmp` doesn't exists in windows and `mkstemp` expects that all path
given already exists (besides the file it will create), causing a fail:
```
../src/tests/eet/eet_test_image.c:93:F:Eet Image:eet_test_image_normal:0: Failure '-1 == (tmpfd = mkstemp(file))' occurred
../src/tests/eet/eet_test_image.c:459:F:Eet Image:eet_test_image_small:0: Failure '-1 == (tmpfd = mkstemp(file))' occurred
```
Here this is solved by using [eina_environment_tmp_get()](https://github.com/expertisesolutions/efl/blob/561410306ee0d00763af2b5d2d0f360de782270c/src/lib/eina/eina_util.c#L97),
which is an architecture-independent function that returns the path to
the temporary directory, as the base of the path.

`eet_close` is a postponed close so when arriving at `unlink` the file
may not be closed yet. On linux it isn't a problem, as when `unlink`ing an
open file it returns `0` and waits until all references to the file to
be closed for its removal.
But on windows you while `unlink`ing an open file `-1` is returnged,  causing a fail:
```
../src/tests/eet/eet_test_image.c:438:F:Eet Image:eet_test_image_normal:0: Failure 'unlink(file) != 0' occurred
../src/tests/eet/eet_test_image.c:497:F:Eet Image:eet_test_image_small:0: Failure 'unlink(file) != 0' occurred
```
With `unlink` setting `errno` to `13`(`strerror(13)` = `Permission denied`).

Here this is solved by forcing a cache clean (`eet_clearcache()` after
closing the file and thus assuring that `unlink` will return `0` on
every architecture.

Note that here for forcing the file to be close both functions are
needed, because `eet_clearcache` needs that the file is marked to be
closed by `eet_close` to then remove it from memory.