Eet: Resolve eet_test_file test case (#391)
Like `eet_test_cache`, this test case had 2 problems: creating a
temporary file and closing it.

About creating the file: `/tmp` doesn't exist in windows and `mkstemp`
expects that all path given already exists (besides the file it will
create), causing a fail:
```
../src/tests/eet/eet_test_file.c:34:F:Eet File:eet_test_file_simple_write:0: Failure '-1 == (tmpfd = mkste mp(file))' occurred
../src/tests/eet/eet_test_file.c:173:F:Eet File:eet_test_file_data:0: Failure '-1 == (tmpfd = mkstemp(file))' occurred
../src/tests/eet/eet_test_file.c:326:F:Eet File:eet_test_file_data_dump:0: Failure '-1 == (tmpfd = mkstemp(file))' occurred
../src/tests/eet/eet_test_file.c:423:F:Eet File:eet_test_file_fp:0: Failure '-1 == (tmpfd = mkstemp(file))' occurred
```
Here this is solved by using `eina_file_mkstemp` that creates a
temporary file at a temporary directory obtained with
`eina_envitonment_tmp` which is an architecture-independent function.

About closing it: `eet_close` is a postponed close, it doesn't immediately close
files opened as `write and read` nor `write` unless it has non saved changes nor
immediately closes files opened as `read` files. So when arriving at `unlink`
the file may not be closed yet.
On linux it isn't a problem, as when `unlink`ing an open file it returns `0` and
waits until all references to the file to be closed for its removal.
But on windows while `unlink`ing an open file `-1` is returnged,  causing a fail
at:
```
../src/tests/eet/eet_test_file.c:124:F:Eet File:eet_test_file_simple_write:0: Failure 'unlink(tmpfile) != 0' occurred
../src/tests/eet/eet_test_file.c:282:P:Eet File:eet_test_file_data:0: Passed
../src/tests/eet/eet_test_file.c:380:F:Eet File:eet_test_file_data_dump:0: Failure 'unlink(tmpfile) != 0' occurred
../src/tests/eet/eet_test_file.c:453:P:Eet File:eet_test_file_fp:0: Passed
```
With `unlink` setting `errno` to `13`(`strerror(13)` = `Permission denied`).

This is solved using `eet_clearcache` which forces the closure of every file (in
case the file has been altered it flushes it before closing).

Note that here `eet_test_file_data` and `eet_test_file_fp` have no problem to
`unlink` the file, this happens because both tests had unsaved changes at the
file while calling `eet_close` which then flushes those changes and really 
closes it.

Note that this solves the problem for *this* test, but it isn't the best
solution, an user still needs to know where his program will run to know if they
need to force file closure before  `unlink`.