ecore_file: Resolve  ecore_test_ecore_file_operations
Work toward  #397

Failing initialy with:
```
../src/tests/ecore/ecore_test_ecore_file.c:186:F:Ecore_File:ecore_test_ecore_file_operations:0: Failure 'res != EINA_TRUE' occurred
```

This happens because `UCRT`'s `remove` only works on files, it doesn't
call `rmdir` in case the specified path is a directory
instead of a file like in Linux.

So this patch changes `ecore_file_remove`: now it verifies if the given
path is a directory and if so calls the correct function
(`ecore_file_rmdir`).

Next, there are some file permission related errors, this happens
because Windows only has 2 file permissions:
- `_S_IREAD ==  0o400` which means a read-only file;
- `_S_IWRITE == 0o200` which means a file with write permission;
- `_S_IWRITE | _S_IREAD == 0o600` which is equivalent to `_S_IWRITE`;
Because of this:
- every file is readable, thus every execution of `ecore_file_can_read`
  return `EINA_TRUE`;
- the concept of an execute permission as in Linux doesn't exist.  On
  Windows only `.exe` files are executables, even `.bat` are not
  considered here (as in `GetBinaryTypeA` function).  `evil` has a
  bypass for the execute permission: `X_OK` (test for execute
  permission) actually returns `F_OK` (test for existence), so as for
  the concept of execute permission as in Linux, every file is
  considered executable because they all exist. Thus, every call to
  `ecore_file_can_exec` return `EINA_TRUE`;
- a file without write permission can not be removed thus making calls
  to `ecore_file_remove` to files without `_S_IWRITE` return
  `EINA_FALSE` with internal `remove` rising `EACESS`.

So here every one of those calls that expects a fail is bypassed. It
should be brought to further discussion at EFL latter;

After this:
```
../src/tests/ecore/ecore_test_ecore_file.c:384:P:Ecore_File:ecore_test_ecore_file_operations:0: Passed
```