ecore: Resolve  ecore_test_file
Work toward  #397

Failing initialy with:
```
../src/tests/ecore/ecore_test_ecore_file.c:186:F:Ecore_File:ecore_test_ecore_file_operations:0: Failure 'res != EINA_TRUE' occurred
```

This happens because `UCRT`'s `remove` only works on files, it doesn't
call `rmdir` in case the specified path is a directory
instead of a file like in Linux.

So this patch changes `ecore_file_remove`: now it verifies if the given
path is a directory and if so calls the correct function
(`ecore_file_rmdir`).

Next error:
```
../src/tests/ecore/ecore_test_ecore_file.c:251:F:Ecore_File:ecore_test_ecore_file_operations:0: Failure 'ecore_file_can_exec(dest_file) != EINA_FALSE' occurred
```

Simillar erros happens after this as in false positives and false
negatives.

The error here is that Windows don't have the concept of an `execute
permission` as in Linux.  On Windows only `.exe` files are executables,
even `.bat` are not considered here (as in `GetBinaryTypeA` function).

`evil` has a bypass for the execute permission: `X_OK` (test for execute
permission) actually returns `F_OK` (test for existence), so for the
concept of executable permission as in Linux, every file is considered
executable because they all exist.

As this concept doesn't exist this PR bypass this tests, to bring this
for futher discussion at EFL latter.

Next was:
```
../src/tests/ecore/ecore_test_ecore_file.c:326:F:Ecore_File:ecore_test_ecore_file_operations:0: Failure 'ecore_file_remove(src_file) != EINA_TRUE' occurred
```
This happens because Windows `remove` return `EACESS` while removing a
read-only file, but Linux works fine with it.

So this ignores the test case where the file is created as read-only as
it couldn't happen as expected in a Linux system.