Evas: Fix evas modules name in native-windows builds
On native-windows we don't have the `MODULE_ARCH` at the path of
modules, so `evas` we don't need to search there;
Another change at the name is that on native-windows builds we want to
use the final name of a modules instead of it intermediary one.

So this changes module names for its final version (`module`) when on
native windows builds. To achieve this, it creates a
`sys_native_windows` variable, which is only true when using `msvc` or
`clang-cl` while in a `windows` host.

Original:
 - 2f3c8295db3e1fa9c62462dbeb6057085e19f87e
 - 426386578bb6f4e045474dbf1d94a502f29ff18b

Co-authored-by: Felipe Magno de Almeida <felipe@expertise.dev>