elm: Fix how elm_config was created
When there is no `elm_config` it was randomly created with
`eina_file_mkstemp` as `<TEMP DIR>/tmp/elementary_config`, which is a
problem because internally `evil` calls `open` to create the file, but
it cannot create a dir. Thus failing at elm initialization.

This was tested at `examples/elementary/label_example_01.c` that was
failing without running anything of itself.
Now it runs to the end, but failing on almost every other call.