elm: Change module names
On Windows, we want to use the final name of modules `module.dll`
instead of its real name like in Linux.

Test Plan
=========
This was tested with `example/elementary/bg_example_0[1,2,3].exe` but it is
expected that with this, there should be no error about failing to load
`prefs_iface` module at elm examples.