win32: Fix wait win32 objects with 64 or more objects
Code was using MsgWaitForMultipleObjects in threads where there were
no Windows, and the main thread used WaitForMultipleObjects, which
caused timeout even if input messages had to be handled.

Also, the quit thread also waited for all threads, which meant waiting
for timeout.

The fix consisted in using MsgWaitForMultipleObjects in the main
thread instead and creating a quit event that made
WaitForMultipleObjects return when a signalled event or message was found.