win32: fix wrong types being used for .cs generation
A variable in .eo with the type (u)long was resulting in an (u)int type in the
generated build/src/tests/efl_mono/dummy_test_object.eo.cs

Sincerly I don't know how it is working even with the lines commented.

Still, there is some issue with (u)llong typed variables that are being
converted to (u)long and being truncated with 32bits.